backend/requirements.txt
# -------------------------------
# FRONTEND (Streamlit)
# -------------------------------
streamlit
plotly

# -------------------------------
# BACKEND (FastAPI)
# -------------------------------
fastapi
uvicorn

# -------------------------------
# DATABASE
# -------------------------------
sqlalchemy
pydantic
python-multipart
pytesseract

# -------------------------------
# MACHINE LEARNING CORE
# -------------------------------
numpy
pandas
scikit-learn==1.6.1
joblib

# -------------------------------
# EMBEDDINGS (SBERT MiniLM local)
# -------------------------------
transformers==4.30.2
torch

# -------------------------------
# OCR (Receipts auto-fill)
# -------------------------------
easyocr
Pillow